+++ ChangeLog.md
@@ -1,5 +1,6 @@
 0.3.0 [2020-08-06]
 
+ * Add progress bar
  * Add --print-groups option to represent
  * Only consider dates if --max-day-sep argument is given (no default, but still sort by date)
  * Do not parse states unless --same-state is given