+++ ChangeLog.md
@@ -1,8 +1,11 @@
 0.3.0 [2020-08-06]
 
+Changes to represent:
+ * Add --print-groups option
  * Add progress bar
- * Add --print-groups option to represent
- * Only consider dates if --max-day-sep argument is given (no default, but still sort by date)
+ * Only consider dates if --max-day-sep argument is given (no default)
+ * If --max-day-sep is given, keep the latest entry from each group, otherwise
+   keep first alphabetically
  * Do not parse states unless --same-state is given 
 
 0.2.0 [2020-01-06]